Move Scenario folder to pages
