+++ src/client/pages/scenario/ScenarioComponent.tsx
