+++ src/client/pages/scenario/ScenarioInfo.tsx
