+++ src/client/pages/scenario/getAllScenarios.ts
