+++ src/client/pages/scenario/getScenarioDescription.ts
